refactor(server): finish audit directory module split
