+++ nodelite-server/src/audit/mod.rs
@@ -1,0 +1,10 @@
+//! Security audit log module wiring and re-exports.
+
+mod log;
+mod query;
+mod storage;
+mod types;
+mod writer;
+
+pub(crate) use self::log::AuditLog;
+pub use self::types::{AuditEvent, AuditEventType, AuditLogError, AuditQuery, NewAuditEvent};
